refactor: initial folder structure for hexagonal arch refactoring
