+++ src/main/java/pauloh/main/adapter/dto/book/BookStockDto.java
@@ -1,0 +1,6 @@
+package pauloh.main.adapter.dto.book;
+
+import java.util.UUID;
+
+public record BookStockDto(UUID bookId, Integer currentStock) {
+}
