+++ src/main/java/pauloh/main/adapter/dto/book/CreateBookDto.java
@@ -1,0 +1,7 @@
+package pauloh.main.adapter.dto.book;
+
+import java.util.UUID;
+
+public record CreateBookDto(UUID userID, String isbn, String title, String author) {
+
+}
