+++ src/main/java/pauloh/main/adapter/dto/book/UpdateBookDto.java
@@ -1,0 +1,7 @@
+package pauloh.main.adapter.dto.book;
+
+import java.util.UUID;
+
+public record UpdateBookDto(UUID bookId, String isbn, String title, String author) {
+
+}
