+++ src/main/java/pauloh/main/adapter/dto/loan/CreateLoanDto.java
@@ -1,0 +1,7 @@
+package pauloh.main.adapter.dto.loan;
+
+import java.util.UUID;
+
+public record CreateLoanDto(UUID userId, UUID bookId) {
+
+}
