+++ src/main/java/pauloh/main/adapter/dto/loan/LoanResponseDto.java
@@ -1,0 +1,8 @@
+package pauloh.main.adapter.dto.loan;
+
+import java.time.Instant;
+import java.util.UUID;
+
+public record LoanResponseDto(UUID id, Instant takenAt, Instant expectedDevolution, Instant devolvedAt) {
+
+}
