+++ src/main/java/pauloh/main/adapter/dto/loan/LoanedBooksWithUsersDto.java
@@ -1,0 +1,5 @@
+package pauloh.main.adapter.dto.loan;
+
+public record LoanedBooksWithUsersDto(String bookTitle, String bookAuthor, String userName, String userEmail) {
+
+}
