+++ src/main/java/pauloh/main/adapter/dto/user/CreateUserDto.java
@@ -1,0 +1,5 @@
+package pauloh.main.adapter.dto.user;
+
+public record CreateUserDto(String name, String email) {
+
+}
