+++ src/main/java/pauloh/main/adapter/dto/user/UserResponseDTO.java
@@ -1,0 +1,7 @@
+package pauloh.main.adapter.dto.user;
+
+import java.util.UUID;
+
+public record UserResponseDTO(UUID id, String name, String email) {
+
+}
