--- src/main/java/pauloh/main/dto/book/BookResponseDto.java
@@ -1,7 +1,0 @@
-package pauloh.main.dto.book;
-
-import java.util.UUID;
-
-public record BookResponseDto(UUID id, String isbn, String title, String author, Integer stock) {
-
-}
