--- src/main/java/pauloh/main/dto/book/BookStockDto.java
@@ -1,6 +1,0 @@
-package pauloh.main.dto.book;
-
-import java.util.UUID;
-
-public record BookStockDto(UUID bookId, Integer currentStock) {
-}
