--- src/main/java/pauloh/main/dto/book/CreateBookDto.java
@@ -1,7 +1,0 @@
-package pauloh.main.dto.book;
-
-import java.util.UUID;
-
-public record CreateBookDto(UUID userID, String isbn, String title, String author) {
-
-}
