--- src/main/java/pauloh/main/dto/book/UpdateBookDto.java
@@ -1,7 +1,0 @@
-package pauloh.main.dto.book;
-
-import java.util.UUID;
-
-public record UpdateBookDto(UUID bookId, String isbn, String title, String author) {
-
-}
