--- src/main/java/pauloh/main/dto/loan/CreateLoanDto.java
@@ -1,7 +1,0 @@
-package pauloh.main.dto.loan;
-
-import java.util.UUID;
-
-public record CreateLoanDto(UUID userId, UUID bookId) {
-
-}
