--- src/main/java/pauloh/main/dto/loan/LoanResponseDto.java
@@ -1,8 +1,0 @@
-package pauloh.main.dto.loan;
-
-import java.time.Instant;
-import java.util.UUID;
-
-public record LoanResponseDto(UUID id, Instant takenAt, Instant expectedDevolution, Instant devolvedAt) {
-
-}
