--- src/main/java/pauloh/main/dto/loan/LoanedBooksWithUsersDto.java
@@ -1,5 +1,0 @@
-package pauloh.main.dto.loan;
-
-public record LoanedBooksWithUsersDto(String bookTitle, String bookAuthor, String userName, String userEmail) {
-
-}
