--- src/main/java/pauloh/main/dto/user/CreateUserDto.java
@@ -1,5 +1,0 @@
-package pauloh.main.dto.user;
-
-public record CreateUserDto(String name, String email) {
-
-}
