--- src/main/java/pauloh/main/dto/user/UserResponseDTO.java
@@ -1,7 +1,0 @@
-package pauloh.main.dto.user;
-
-import java.util.UUID;
-
-public record UserResponseDTO(UUID id, String name, String email) {
-
-}
