+++ src/main/java/pauloh/main/port/input/UserInputPort.java
@@ -1,0 +1,5 @@
+package pauloh.main.port.input;
+
+public interface UserInputPort {
+  // TODO
+}
