+++ src/main/java/pauloh/main/port/output/UserOutputPort.java
@@ -1,0 +1,5 @@
+package pauloh.main.port.output;
+
+public interface UserOutputPort {
+  // TODO
+}
